manage pods in unknown state
This PR adds an additonal func on `operator` that is called
`managePodHealth` which helps to look at the status of the pod from
the pod template spec. If the satus is `PodUnknown`, then delete
action is called on the pod.

This new functionality is added to the reconcile logic loop
and the operator reconciles and kills pods that are stuck in
a unknown states every reconciliation loop.

`Note:` This PR needs tests, and PR would be updated.