Implements support for assigning multiple partitions per deployment.
Add new optional field `numPartitionsPerInstance`to the consumerSpec to instruct
controller to assign multiple partitions per deployment. Partition annotation contains
a comma-separated list of partitions. It still can parse previous generations of the deployments where only one partition is assigned.

New annotation `consumer-id` is added to the deployment to indicate the ord of the consumer.

`Estimate` function of the `Predictor` interface is now receives a list of partitions instead of a single one and returns the total amount of resources needed to consume from those partitions.