Fix state counters and metrics reporting
Previously each deployment was responsible for one partition, and all counters were built around partitions.
Now deployment could manage multiple partitions, which means we need to operate on deployment counters. This also require track consumerIds using maps instead of lists.

This change adds `Redundant` field to the status object to indicate number of deployments that konsumerator is going to delete. This is important because the number of deployments could increase and decrease more often now.

Add few more log lines to indicate important decisions, like which instances are going to be deleted or created.