Resync deployment metrics on reconcile.
Metrics are wrong in case of changing the number of partitions per consumer,
because controller doesn't remove metrics from the old deployments. The only
way to fix this is by deleting the controller pod, but in this case controller
doesn't reexport metrics from the deployments that are already running.

With this change controller will sync deployment metrics on each reconcile.