Fix premature saturation
There is a bug in status-change state machine. In case of
a drastic surge of production rate (2x-3x) which results in
resource estimations being higher than maximum allowed,
konsumerator sets the status to "saturated" without actually
scaling the consumer to the max. Fix this and add a few tests.