Fix bug in global limiter
GlobalLimiter returns false positive when only one of the resource pools is being exhausted. This leads to the instance not being scaled up when it's needed.

For example, when memory pool is empty (due to 1-1 allocation), instances will only be scaled up (cpu-wise) when some instance isn't running, otherwise it will be locked with amount of CPUs set to it during creation.