Implement fallback strategy when there is no metrics
Add new field to the spec - `fallbackStrategy` that controlls what
amount of resources should be assigned to the instance when it's
impossible to run estimations due to missing metrics.

Only two modes are implemented at the moment - min and max, with default
being min (to preserve the existing behaviour).

Fallback strategy assigns min/max value based on either of 2 inputs:

* if there are more than half of the instances are running (and more
  than 10 in total), find min/max among those instances and use it as
  fallback value for new estimations.

* otherwise, use min/max from the container policy set in the spec:
  ```
	containerPolicies:
	- containerName: name
		minAllowed:
			cpu: 3
			memory: 12Gi
		maxAllowed:
			cpu: 8
			memory: 12Gi
  ```