Fix state machine when instances stuck in SATURATED
There is a case when we need to explicitely reset SATURATED state.

If estimation returns `eq` or `gt` and some amount of resources got
cut by any of the limiters, instances will continue to run in SATURATED
state, even when there is no lag. Fix this by explicitely cleaning up
"saturation".